% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_metadata}
\alias{add_metadata}
\title{Add DX_bl, AGE, SEX, ... data}
\usage{
add_metadata(mt, other_info, baseline = TRUE)
}
\arguments{
\item{mt}{Main dataset.}

\item{other_info}{Additional information containing DX_bl, AGE, SEX, ... data.}

\item{baseline}{When TRUE, the metadata is filter by baseline.
otherwise no filter. TRUE is set as default.}
}
\value{
Data frame with added DX_bl, AGE, SEX, ... information.
}
\description{
this \code{f()} adds DX_bl, AGE, SEX, ... information to the dataset
based on sample identifiers.
}
\examples{
\dontrun{
updated_data <- add_metadata(mt, other_info)
}
}
