% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADNIMERGE.R
\docType{class}
\name{ADNIMERGE-class}
\alias{ADNIMERGE-class}
\alias{ADNIMERGE}
\title{The ADNIMERGE class}
\value{
An ADNIMERGE class object.
}
\description{
The ADNIMERGE object is a representation of the ADNIMERGE dataset,
containing demographic, clinical, biomarker, imaging, cognitive, and
other relevant data.
}
\section{Slots}{

\describe{
\item{\code{demographic_data}}{a data.frame or tbl_df of
\link{demographic_data}.}

\item{\code{baseline_clinical_data}}{a data.frame or tbl_df of
\link{baseline_clinical_data}.}

\item{\code{follow_up_clinical_data}}{a data.frame or tbl_df of
\link{follow_up_clinical_data}.}

\item{\code{biomarker_data}}{a data.frame or tbl_df of
\link{biomarker_data}.}

\item{\code{imaging_data}}{a data.frame or tbl_df of \link{imaging_data}.}

\item{\code{cognitive_composite_data}}{a data.frame or tbl_df of
\link{cognitive_composite_data}.}

\item{\code{ecog_scores_data}}{a data.frame or tbl_df of
\link{ecog_scores_data}.}

\item{\code{RID_to_PTID}}{a data.frame or tbl_df linking \link{RID_to_PTID}.}

\item{\code{logs}}{Log information of data processing steps.}

\item{\code{misc_data}}{...}
}}

\seealso{
\link{ADNIMERGE}
}
