% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_normalization.R
\name{batch_normalization}
\alias{batch_normalization}
\title{batch normalization for a tibble}
\usage{
batch_normalization(
  tbl,
  try = FALSE,
  batch_name = "Plate Bar Code",
  verbose = TRUE
)
}
\arguments{
\item{tbl}{is a tibble object. The first two columns are "ID" and "Batch", the remaining columns for batch normalization}

\item{try}{test the function for the first 20 columns.}

\item{verbose}{print log information.}
}
\value{
A data.frame object after normalization.
}
\description{
Normalization data by the median value of each batch in a tibble
}
