% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_composite_data.R
\docType{data}
\name{cognitive_composite_data}
\alias{cognitive_composite_data}
\title{Cognitive composite data}
\format{
A data frame with rows and variables as described in the ADNI
documentation.
\describe{
\item{RID}{Participant roster ID}
\item{mPACCdigit}{ADNI modified Preclinical Alzheimer's Cognitive
Composite (PACC) with Digit Symbol Substitution}
\item{mPACCtrailsB}{ADNI modified Preclinical Alzheimer's Cognitive
Composite (PACC) with Trails B}
}
}
\source{
Alzheimer's Disease Neuroimaging Initiative (ADNI)
}
\usage{
cognitive_composite_data
}
\description{
This dataset contains cognitive composite scores from the Alzheimer's
Disease Neuroimaging Initiative (ADNI).
}
\keyword{datasets}
