% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_data.R
\docType{data}
\name{demographic_data}
\alias{demographic_data}
\title{Demographic data}
\format{
A data frame with rows and variables as described in the ADNI
documentation.
\describe{
\item{RID}{Participant roster ID}
\item{AGE}{Age}
\item{PTGENDER}{Sex}
\item{PTEDUCAT}{Education}
\item{PTETHCAT}{Ethnicity}
\item{PTRACCAT}{Race}
\item{PTMARRY}{Marital status}
}
}
\source{
Alzheimer's Disease Neuroimaging Initiative (ADNI)
}
\usage{
demographic_data
}
\description{
This dataset contains demographic information from the Alzheimer's Disease
Neuroimaging Initiative (ADNI).
}
\keyword{datasets}
