% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fncols}
\alias{fncols}
\title{Add missing columns to data frame}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{Data frame to check.}

\item{cname}{Vector of column names to add if missing.}
}
\value{
A data frame with added columns (if missing).
}
\description{
this \code{f()} ensures a data frame contains specified columns,
adding them as \code{NA} if missing.
}
\examples{
\dontrun{
data <- fncols(data, c("new_col1", "new_col2"))
}
}
