% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging_data.R
\docType{data}
\name{imaging_data}
\alias{imaging_data}
\title{Imaging data}
\format{
A data frame with rows and variables as described in the ADNI
documentation.
\describe{
\item{RID}{Participant roster ID}
\item{FLDSTRENG}{MRI Field Strength}
\item{FSVERSION}{FreeSurfer Software Version}
\item{Ventricles}{UCSF Ventricles}
\item{Hippocampus}{UCSF Hippocampus}
\item{WholeBrain}{UCSF WholeBrain}
\item{Entorhinal}{UCSF Entorhinal}
\item{Fusiform}{UCSF Fusiform}
\item{MidTemp}{UCSF Med Temp}
\item{ICV}{UCSF ICV}
}
}
\source{
Alzheimer's Disease Neuroimaging Initiative (ADNI)
}
\usage{
imaging_data
}
\description{
This dataset contains imaging measurements from the Alzheimer's Disease
Neuroimaging Initiative (ADNI).
}
\keyword{datasets}
