% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_save_roc.R
\name{plot_and_save_roc}
\alias{plot_and_save_roc}
\title{Plot and Save a ROC Curve with AUC Annotation}
\usage{
plot_and_save_roc(
  roc_obj,
  filename = "roc_curve.png",
  title = "ROC Curve",
  width = 5,
  height = 5,
  line_color = "blue",
  auc_digits = 2,
  auc_position = c(0.6, 0.1),
  smooth = FALSE,
  ci = FALSE,
  ci_level = 0.95,
  ci_size = 0.5,
  ci_alpha = 0.3,
  ci_auc = TRUE
)
}
\arguments{
\item{roc_obj}{A \code{pROC::roc} object.}

\item{filename}{Output filename for the saved plot (e.g., "roc_curve.png").}

\item{title}{Title of the plot.}

\item{width}{Width of the saved image in inches.}

\item{height}{Height of the saved image in inches.}

\item{line_color}{Color of the ROC curve line.}

\item{auc_digits}{Number of digits to round the AUC value.}

\item{auc_position}{A numeric vector of length 2 indicating the (x, y)
position of the AUC label (in plot coordinates).}

\item{smooth}{Logical; if \code{TRUE}, the ROC curve will be smoothed (unless}

\item{ci}{Logical; if TRUE, plots confidence interval around the ROC curve.}

\item{ci_level}{Confidence level for interval (default 0.95).}

\item{ci_size}{Line size for confidence interval ribbon.}

\item{ci_alpha}{Transparency for confidence interval ribbon.}

\item{ci_auc}{if TRUE, add confidence interval for AUC
already).}
}
\value{
The ggplot object (invisible).
}
\description{
Generates and saves a ggplot2-based ROC curve from a \code{pROC::roc} object,
with the AUC displayed on the plot.
}
\examples{
\dontrun{
roc_obj <- pROC::roc(
    response = test$Allgr, predictor = top_probs, levels = c("CN", "AD")
)
plot_and_save_roc(roc_obj, "roc_curve_top.png", title = "Top 5 Metabolites")
}
}
