% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RID_to_PTID.R
\docType{data}
\name{RID_to_PTID}
\alias{RID_to_PTID}
\title{RID - PTID}
\format{
A data frame with rows and variables as described in the ADNI
documentation.
\describe{
\item{RID}{Participant roster ID}
\item{PTID}{Original study protocol}
}
}
\source{
Alzheimer's Disease Neuroimaging Initiative (ADNI)
}
\usage{
RID_to_PTID
}
\description{
This dataset contains links \code{RIDs} to \code{PTIDS} from the Alzheimer's Disease
Neuroimaging Initiative (ADNI).
}
\keyword{datasets}
