% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{student_fdr}
\alias{student_fdr}
\title{student_fdr}
\usage{
student_fdr(
  start,
  end,
  group_1,
  group_2,
  log,
  base,
  comparison,
  group_1_original = NULL,
  group_2_original = NULL
)
}
\arguments{
\item{start}{Numeric, the starting column index of your matrix
or data frame.}

\item{end}{Numeric, the ending column index of your matrix or data frame.}

\item{group_1}{Data frame or matrix for the first group.}

\item{group_2}{Data frame or matrix for the second group.}

\item{log}{Logical, TRUE if your data are log-transformed.}

\item{base}{Numeric, base of the log transformation}

\item{comparison}{Character, name of the comparison for the output file.}

\item{group_1_original}{first group group original data set}

\item{group_2_original}{second group original data set
(ignored if log = FALSE).}
}
\value{
the function returns a data frame with mean, standard deviation,
p-value, fold-change, and FDR-adjusted p-value for each variable.
}
\description{
Calculate a Student test and the foldchange
}
