% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_renaming.R
\name{transform_colnames}
\alias{transform_colnames}
\title{Transform column names of the FIA dataset}
\usage{
transform_colnames(old_name)
}
\arguments{
\item{old_name}{string parameter representing the old column name.}
}
\value{
string parameter with the transformed column name.
}
\description{
This function applies a series of transformations to the column names
of a data frame. It performs specific replacements for common column name
patterns and handles special cases for various lipid and compound names
(e.g., Ceramides, HexCer, SM, PC, etc.).
}
\details{
The function handles the following transformations:
\itemize{
\item Special cases for specific column names like "PTID" to "RID",
"Customer_Sample_Identification" to "Customer Sample Identification", etc.
\item Replacement of underscores with colons for compound names
like "C3_1" to "C3:1".
\item Transformation of OH/DC notation compounds
(e.g., "\emph{OH}_" to "-OH (", "\emph{DC}_" to "-DC (").
\item Special formatting for Ceramide, HexCer, Hex2Cer, and Hex3Cer names.
\item SM, lysoPC, and PC aa/ae transformations.
\item Support for handling DG and TG names, including modifications
to handle isotopic forms.
\item Reformatting of CE names.
}
}
\examples{
# e.g. of applying the transformation function to a set of column names:
old_names <- c("PTID", "Customer_Sample_Identification", "C3_1",
    "Cer_d18:1_18:1")
# old_names <- colnames(FIA_Metabo)
new_names <- sapply(old_names, transform_colnames, USE.NAMES = FALSE)
print(new_names)

}
