% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits.R
\docType{data}
\name{visits}
\alias{visits}
\title{Baseline clinical data}
\format{
A data frame with rows and variables as described in the ADNI
documentation.
\describe{
\item{Phase}{ADNI stady phase}
\item{ID}{Unique identifier}
\item{VISCODE}{Visit code}
\item{VISNAME}{Visit name}
\item{VISORDER}{Visit order}
}
}
\source{
Alzheimer's Disease Neuroimaging Initiative (ADNI)
}
\usage{
visits
}
\description{
This dataset contains visits details from the Alzheimer's
Disease Neuroimaging Initiative (ADNI).
}
\keyword{datasets}
