% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{welch_fdr}
\alias{welch_fdr}
\title{welch_fdr}
\usage{
welch_fdr(start, end, group_1, group_2, log, base, comparison)
}
\arguments{
\item{start}{Numeric, the starting column index of your matrix
or data frame.}

\item{end}{Numeric, the ending column index of your matrix or data frame.}

\item{group_1}{Data frame or matrix for the first group.}

\item{group_2}{Data frame or matrix for the second group.}

\item{log}{Logical, TRUE if your data are log-transformed.}

\item{base}{Numeric, base of the log transformation
(ignored if log = FALSE).}

\item{comparison}{Character, name of the comparison for the output file.}
}
\value{
The function returns a data frame with mean, standard deviation,
p-value, fold-change, and FDR-adjusted p-value for each variable.
}
\description{
Calculate a Welch t-test and the fold-change
}
